requirements.txt
pandas
numpy
scikit-learn
fastapi
uvicorn
pydantic
joblib
matplotlib
djangorestframework 
djangorestframework-simplejwt
geopy

